ml_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3
mysql-connector-python==8.2.0
python-multipart==0.0.6
pydantic==2.5.0
joblib==1.3.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.2
faker==20.1.0
textblob==0.17.1
cryptography>=41.0.0
requests>=2.31.0
gunicorn>=21.2.0
